# swagger_server/models/cities_hint_model.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class CitiesHintModel(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, name: str=None):  # noqa: E501
        """CitiesHintModel - a model defined in Swagger

        :param id: The id of this CitiesHintModel.  # noqa: E501
        :type id: int
        :param name: The name of this CitiesHintModel.  # noqa: E501
        :type name: str
        """
        self.swagger_types = {
            'id': int,
            'name': str
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name'
        }
        self._id = id
        self._name = name

    @classmethod
    def from_dict(cls, dikt) -> 'CitiesHintModel':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CitiesHintModel of this CitiesHintModel.  # noqa: E501
        :rtype: CitiesHintModel
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this CitiesHintModel.


        :return: The id of this CitiesHintModel.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this CitiesHintModel.


        :param id: The id of this CitiesHintModel.
        :type id: int
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this CitiesHintModel.


        :return: The name of this CitiesHintModel.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this CitiesHintModel.


        :param name: The name of this CitiesHintModel.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name