# swagger_server/models/error_response_model.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class ErrorResponseModel(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, code: int=None, message: str=None):  # noqa: E501
        """ErrorResponseModel - a model defined in Swagger

        :param code: The code of this ErrorResponseModel.  # noqa: E501
        :type code: int
        :param message: The message of this ErrorResponseModel.  # noqa: E501
        :type message: str
        """
        self.swagger_types = {
            'code': int,
            'message': str
        }

        self.attribute_map = {
            'code': 'code',
            'message': 'message'
        }
        self._code = code
        self._message = message

    @classmethod
    def from_dict(cls, dikt) -> 'ErrorResponseModel':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ErrorResponseModel of this ErrorResponseModel.  # noqa: E501
        :rtype: ErrorResponseModel
        """
        return util.deserialize_model(dikt, cls)

    @property
    def code(self) -> int:
        """Gets the code of this ErrorResponseModel.


        :return: The code of this ErrorResponseModel.
        :rtype: int
        """
        return self._code

    @code.setter
    def code(self, code: int):
        """Sets the code of this ErrorResponseModel.


        :param code: The code of this ErrorResponseModel.
        :type code: int
        """
        if code is None:
            raise ValueError("Invalid value for `code`, must not be `None`")  # noqa: E501

        self._code = code

    @property
    def message(self) -> str:
        """Gets the message of this ErrorResponseModel.


        :return: The message of this ErrorResponseModel.
        :rtype: str
        """
        return self._message

    @message.setter
    def message(self, message: str):
        """Sets the message of this ErrorResponseModel.


        :param message: The message of this ErrorResponseModel.
        :type message: str
        """
        if message is None:
            raise ValueError("Invalid value for `message`, must not be `None`")  # noqa: E501

        self._message = message